src/main/webapp/app/entities/spaces/update/spaces-update.component.spec.ts
jest.mock('@angular/router');

import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { of, Subject } from 'rxjs';

import { SpacesService } from '../service/spaces.service';
import { ISpaces, Spaces } from '../spaces.model';

import { IUser } from 'app/entities/user/user.model';
import { UserService } from 'app/entities/user/user.service';

import { SpacesUpdateComponent } from './spaces-update.component';

describe('Component Tests', () => {
  describe('Spaces Management Update Component', () => {
    let comp: SpacesUpdateComponent;
    let fixture: ComponentFixture<SpacesUpdateComponent>;
    let activatedRoute: ActivatedRoute;
    let spacesService: SpacesService;
    let userService: UserService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [SpacesUpdateComponent],
        providers: [FormBuilder, ActivatedRoute],
      })
        .overrideTemplate(SpacesUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(SpacesUpdateComponent);
      activatedRoute = TestBed.inject(ActivatedRoute);
      spacesService = TestBed.inject(SpacesService);
      userService = TestBed.inject(UserService);

      comp = fixture.componentInstance;
    });

    describe('ngOnInit', () => {
      it('Should call User query and add missing value', () => {
        const spaces: ISpaces = { id: 456 };
        const user: IUser = { id: 10396 };
        spaces.user = user;

        const userCollection: IUser[] = [{ id: 69687 }];
        jest.spyOn(userService, 'query').mockReturnValue(of(new HttpResponse({ body: userCollection })));
        const additionalUsers = [user];
        const expectedCollection: IUser[] = [...additionalUsers, ...userCollection];
        jest.spyOn(userService, 'addUserToCollectionIfMissing').mockReturnValue(expectedCollection);

        activatedRoute.data = of({ spaces });
        comp.ngOnInit();

        expect(userService.query).toHaveBeenCalled();
        expect(userService.addUserToCollectionIfMissing).toHaveBeenCalledWith(userCollection, ...additionalUsers);
        expect(comp.usersSharedCollection).toEqual(expectedCollection);
      });

      it('Should update editForm', () => {
        const spaces: ISpaces = { id: 456 };
        const user: IUser = { id: 3971 };
        spaces.user = user;

        activatedRoute.data = of({ spaces });
        comp.ngOnInit();

        expect(comp.editForm.value).toEqual(expect.objectContaining(spaces));
        expect(comp.usersSharedCollection).toContain(user);
      });
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Spaces>>();
        const spaces = { id: 123 };
        jest.spyOn(spacesService, 'update').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ spaces });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.next(new HttpResponse({ body: spaces }));
        saveSubject.complete();

        // THEN
        expect(comp.previousState).toHaveBeenCalled();
        expect(spacesService.update).toHaveBeenCalledWith(spaces);
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Spaces>>();
        const spaces = new Spaces();
        jest.spyOn(spacesService, 'create').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ spaces });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.next(new HttpResponse({ body: spaces }));
        saveSubject.complete();

        // THEN
        expect(spacesService.create).toHaveBeenCalledWith(spaces);
        expect(comp.isSaving).toEqual(false);
        expect(comp.previousState).toHaveBeenCalled();
      });

      it('Should set isSaving to false on error', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Spaces>>();
        const spaces = { id: 123 };
        jest.spyOn(spacesService, 'update').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ spaces });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.error('This is an error!');

        // THEN
        expect(spacesService.update).toHaveBeenCalledWith(spaces);
        expect(comp.isSaving).toEqual(false);
        expect(comp.previousState).not.toHaveBeenCalled();
      });
    });

    describe('Tracking relationships identifiers', () => {
      describe('trackUserById', () => {
        it('Should return tracked User primary key', () => {
          const entity = { id: 123 };
          const trackResult = comp.trackUserById(0, entity);
          expect(trackResult).toEqual(entity.id);
        });
      });
    });
  });
});
